Move the insight module into the cli/utils directory
