+++ cli/.gitignore
@@ -2,3 +2,4 @@
 stages/*.py
 build/
 dist/
+insight.json
