+++ cli/utils/insight.py
