+++ cli/utils/insight_tests.py
